added combo bar
- combo is now a float from 1-7

+++ ArcadeJam/Player/Player.cs
@@ -20,7 +20,7 @@ public class Player : Node {
 	Vector2Data Vel { get; set; } = new(new Vector2(0, 0));
 	FloatData moveSpeed = new(1.5f);
 
-	public IntData combo{ get; private set;} = new IntData(1);
+	public FloatData combo{ get; private set;} = new FloatData(1);
 	IntData score;
 	Sprite Sprite { get; set; } = new Sprite(Assets.player);
 	List<Node> collisions = new();
